Fixing issues in auth service

+++ server/event-service/Controllers/OrganizerController.cs
@@ -1,10 +1,13 @@
 ﻿using AutoMapper;
 using event_service.DTOs;
 using event_service.Entities;
+using event_service.Extensions;
 using event_service.Services.IServices;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Configuration;
 using Microsoft.IdentityModel.Tokens;
+using System.Management;
 
 namespace event_service.Controllers
 {
@@ -14,12 +17,80 @@ namespace event_service.Controllers
         private readonly IEventService _eventService;
         private readonly IUserService _userService;
         private readonly IMapper _mapper;
+        private readonly IConfiguration _configuration;
 
-        public OrganizerController(IEventService eventService, IMapper mapper, IUserService userService)
+
+        public OrganizerController(IEventService eventService, IMapper mapper, IUserService userService, IConfiguration configuration)
         {
             _eventService = eventService;
             _mapper = mapper;
             _userService = userService;
+            _configuration = configuration;
+        }
+
+        private bool AuthorizeRole(string role)
+        {
+            // Retrieve JWT token from the request headers
+            var jwtToken = HttpContext.Request.Cookies["jwtToken"];
+            if (string.IsNullOrEmpty(jwtToken))
+            {
+                return false;
+            }
+
+            // Initialize JwtTokenValidator with the issuer, audience, and secret key
+            var Secret = Environment.GetEnvironmentVariable("SECRET");
+            var Issuer = Environment.GetEnvironmentVariable("ISSUER");
+            var Audience = Environment.GetEnvironmentVariable("AUDIENCE");
+            var tokenValidator = new JwtTokenValidator(Issuer, Audience, Secret);
+
+
+            /*var jwtOptions = _configuration.GetSection("ApiSettings:JwtOptions").Get<JwtOptions>();
+            var tokenValidator = new JwtTokenValidator(jwtOptions.Issuer, jwtOptions.Audience, jwtOptions.Secret);*/
+
+
+            // Validate JWT token and extract user roles
+            var roles = tokenValidator.ValidateTokenRole(jwtToken);
+
+            if (roles.Contains(role))
+            {
+                return true;
+            } 
+            else
+            {
+                return false;
+            }
+        }
+
+        private bool AuthorizeID(string ID)
+        {
+            // Retrieve JWT token from the request headers
+            var jwtToken = HttpContext.Request.Cookies["jwtToken"];
+            if (string.IsNullOrEmpty(jwtToken))
+            {
+                return false;
+            }
+
+            // Initialize JwtTokenValidator with the issuer, audience, and secret key
+            var Secret = Environment.GetEnvironmentVariable("SECRET");
+            var Issuer = Environment.GetEnvironmentVariable("ISSUER");
+            var Audience = Environment.GetEnvironmentVariable("AUDIENCE");
+            var tokenValidator = new JwtTokenValidator(Issuer, Audience, Secret);
+
+
+            /*var jwtOptions = _configuration.GetSection("ApiSettings:JwtOptions").Get<JwtOptions>();
+            var tokenValidator = new JwtTokenValidator(jwtOptions.Issuer, jwtOptions.Audience, jwtOptions.Secret);*/
+
+            // Validate JWT token and extract user roles
+            var IDs = tokenValidator.ValidateTokenID(jwtToken);
+
+            if (IDs.Contains(ID))
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
         }
 
         [HttpGet("Organizers")]
@@ -45,19 +116,30 @@ namespace event_service.Controllers
 
 
         [HttpPost("Organizer/{OrganizerId}/Events")]
-        [Authorize(Roles = "ORGANIZER")]
         public async Task<IActionResult> CreateEvent([FromBody] EventReqDto ev, string OrganizerId)
         {
             try
             {
-                if (await _eventService.CreateEvent(ev, OrganizerId))
+                if (AuthorizeID(OrganizerId) && (AuthorizeRole("ORGANIZER") || AuthorizeRole("ADMIN")))
                 {
-                    return Ok(ev);
+                    if (!await _userService.IsOrganizerExists(OrganizerId))
+                        return NotFound("Organizer not found");
+
+                    if (await _eventService.CreateEvent(ev, OrganizerId))
+                    {
+                        return Ok(ev);
+                    }
+                    else
+                    {
+                        return BadRequest("Failed to create event.");
+                    }
                 }
                 else
                 {
-                    return BadRequest("Failed to create event.");
+                    // User does not have the required role
+                    return Unauthorized();
                 }
+
             }
             catch (Exception ex)
             {
@@ -66,20 +148,30 @@ namespace event_service.Controllers
         }
 
         [HttpGet("Organizer/{OrganizerId}/Events")]
-        [Authorize(Roles = "ORGANIZER")]
-        public async Task<IActionResult> GetAllEvents(string OrganizerId)
+        public async Task<IActionResult> GetAllEvents(string OrganizerId, int pageNumber = 1)
         {
             try
             {
-                if (!await _eventService.OrganizerHasEvents(OrganizerId))
-                    return NotFound();
+                if (AuthorizeID(OrganizerId) && (AuthorizeRole("ORGANIZER") || AuthorizeRole("ADMIN")))
+                {
+                    if (!await _userService.IsOrganizerExists(OrganizerId))
+                        return NotFound("Organizer not found");
 
-                var events = _mapper.Map<ICollection<EventsDto>>(await _eventService.GetEventsByOrganizer(OrganizerId));
+                    if (!await _eventService.OrganizerHasEvents(OrganizerId))
+                        return NotFound();
 
-                if (!ModelState.IsValid)
-                    return BadRequest(ModelState);
+                    var events = await _eventService.GetEventsByOrganizer(OrganizerId, pageNumber);
 
-                return Ok(events);
+                    if (!ModelState.IsValid)
+                        return BadRequest(ModelState);
+
+                    return Ok(events);
+                }
+                else
+                {
+                    // User does not have the required role
+                    return Unauthorized();
+                }
             }
             catch (Exception ex)
             {
@@ -88,20 +180,30 @@ namespace event_service.Controllers
         }
 
         [HttpGet("Organizer/{OrganizerId}/Events/{EventId}")]
-        [Authorize(Roles = "ORGANIZER")]
         public async Task<IActionResult> GetEventById(string OrganizerId, int EventId)
         {
             try
             {
-                if (!await _eventService.IsEventExist(EventId))
-                    return NotFound();
+                if (AuthorizeID(OrganizerId) && (AuthorizeRole("ORGANIZER") || AuthorizeRole("ADMIN")))
+                {
+                    if (!await _userService.IsOrganizerExists(OrganizerId))
+                        return NotFound("Organizer not found");
 
-                var _event = _mapper.Map<EventByIdDto>(await _eventService.GetEventById(OrganizerId, EventId));
+                    if (!await _eventService.IsEventExist(EventId))
+                        return NotFound("Event not found");
 
-                if (!ModelState.IsValid)
-                    return BadRequest(ModelState);
+                    var _event = _mapper.Map<EventByIdDto>(await _eventService.GetEventById(OrganizerId, EventId));
 
-                return Ok(_event);
+                    if (!ModelState.IsValid)
+                        return BadRequest(ModelState);
+
+                    return Ok(_event);
+                }
+                else
+                {
+                    // User does not have the required role
+                    return Unauthorized();
+                } 
             }
             catch (Exception ex)
             {
@@ -110,23 +212,33 @@ namespace event_service.Controllers
         }
 
         [HttpDelete("Organizer/{OrganizerId}/Events/{EventId}")]
-        [Authorize(Roles = "ORGANIZER")]
         public async Task<IActionResult> DeleteEvent(string OrganizerId, int EventId)
         {
             try
             {
-                if (!await _eventService.IsEventExist(EventId))
-                    return NotFound();
+                if (AuthorizeID(OrganizerId) && (AuthorizeRole("ORGANIZER") || AuthorizeRole("ADMIN")))
+                {
+                    if (!await _userService.IsOrganizerExists(OrganizerId))
+                        return NotFound("Organizer not found");
 
-                var temp = await _eventService.DeleteEvent(OrganizerId, EventId);
+                    if (!await _eventService.IsEventExist(EventId))
+                        return NotFound("Event not found");
 
-                if (!ModelState.IsValid)
-                    return BadRequest(ModelState);
+                    var temp = await _eventService.DeleteEvent(OrganizerId, EventId);
 
-                if (temp)
-                    return Ok(temp);
-                else 
-                    return BadRequest("Failed to delete event.");
+                    if (!ModelState.IsValid)
+                        return BadRequest(ModelState);
+
+                    if (temp)
+                        return Ok(temp);
+                    else
+                        return BadRequest("Failed to delete event.");
+                }
+                else
+                {
+                    // User does not have the required role
+                    return Unauthorized();
+                }            
             }
             catch (Exception ex)
             {
@@ -135,23 +247,33 @@ namespace event_service.Controllers
         }
         
         [HttpPut("Organizer/{OrganizerId}/Events/{EventId}")]
-        [Authorize(Roles = "ORGANIZER")]
         public async Task<IActionResult> UpdateEvent([FromBody] EventReqDto ev, string OrganizerId, int EventId)
         {
             try
             {
-                if (!await _eventService.IsEventExist(EventId))
-                    return NotFound();
+                if (AuthorizeID(OrganizerId) && (AuthorizeRole("ORGANIZER") || AuthorizeRole("ADMIN")))
+                {
+                    if (!await _userService.IsOrganizerExists(OrganizerId))
+                        return NotFound("Organizer not found");
 
-                var temp = await _eventService.UpdateEvent(ev, OrganizerId, EventId);
+                    if (!await _eventService.IsEventExist(EventId))
+                        return NotFound("Event not found");
 
-                if (!ModelState.IsValid)
-                    return BadRequest(ModelState);
+                    var temp = await _eventService.UpdateEvent(ev, OrganizerId, EventId);
 
-                if (temp)
-                    return Ok(temp);
+                    if (!ModelState.IsValid)
+                        return BadRequest(ModelState);
+
+                    if (temp)
+                        return Ok(temp);
+                    else
+                        return BadRequest("Failed to update event.");
+                }
                 else
-                    return BadRequest("Failed to update event.");
+                {
+                    // User does not have the required role
+                    return Unauthorized();
+                }
             }
             catch (Exception ex)
             {
